Points awarded before voting
With the old rule, points awarded happened at the end of the turn. With our unbiased random system, this means the amount of points awarded would be known throughout the voting process. Not good. Adding an additional step where points are awarded after the rule is convinced, but before it is voted on.